minor modification

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,9 +7,9 @@
  * @action: function pointer
  */
 
-void array_iterator(int *array, size_t size, void (*action)(int))
+void array_iterator(int *array, int size, void (*action)(int))
 {
-	size_t i;
+	int i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
